fix(coding-agent): statically register attach command
